Use MUnit in PathTreeSuite
